Added a bannable trait
