+++ src/Filters/SessionAuth.php
@@ -51,10 +51,18 @@ class SessionAuth implements FilterInterface
 
             // Block inactive users when Email Activation is enabled
             $user = $authenticator->getUser();
+
+            if ($user->isBanned()) {
+                $authenticator->logout();
+
+                return redirect()->to(config('Auth')->logoutRedirect())
+                    ->with('error', lang('Auth.logOutBannedUser'));
+            }
+
             if ($user !== null && ! $user->isActivated()) {
                 $authenticator->logout();
 
-                return redirect()->route('login')
+                return redirect()->to(config('Auth')->logoutRedirect())
                     ->with('error', lang('Auth.activationBlocked'));
             }
