+++ tests/Authorization/AuthorizableTest.php
@@ -326,4 +326,24 @@ final class AuthorizableTest extends DatabaseTestCase
         Locale::setDefault($currentLocale);
         Time::setTestNow();
     }
+
+    public function testBanningUser(): void
+    {
+        $this->assertFalse($this->user->isBanned());
+
+        $this->user->ban();
+
+        $this->assertTrue($this->user->isBanned());
+    }
+
+    public function testUnbanningUser(): void
+    {
+        $this->user->ban();
+
+        $this->assertTrue($this->user->isBanned());
+
+        $this->user->unBan();
+
+        $this->assertFalse($this->user->isBanned());
+    }
 }
